added testnet

+++ Sources/MempoolKit/Models/RecommendedFees.swift
@@ -2,6 +2,6 @@
 import Foundation
 
 public struct RecommendedFees: Codable {
-    public let fastestFee, halfHourFee, hourFee, economyFee, minimumFee: Int
+    public let fastestFee, halfHourFee, hourFee, economyFee, minimumFee: Double
 }
